scripts/migration/requirements.txt
click>=8.1.0
pyyaml>=6.0
psycopg2-binary>=2.9.0
pymysql>=1.1.0
httpx>=0.27.0
rich>=13.0.0
